edit

+++ expenses/views.py
@@ -3,12 +3,33 @@ from django.contrib.auth.decorators import login_required
 from .models import Category, Expense
 from django.contrib import messages
 from django.core.paginator import Paginator
+import json
+from django.http import JsonResponse
+from django.http import HttpResponse
 # Create your views here.
+
+
+def search_expenses(request):
+    if request.method == 'POST':
+        search_str = json.loads(request.body).get('searchText')
+
+        expenses = Expense.objects.filter(
+            amount__istartswith = search_str, owner= request.user) | Expense.objects.filter(
+            date__istartswith = search_str, owner= request.user) | Expense.objects.filter(
+            description__icontains = search_str, owner= request.user) | Expense.objects.filter(
+            category__icontains = search_str, owner= request.user)
+        
+        data = expenses.values()
+        return JsonResponse(list(data), safe= False )
+
+
+
+
 
 @login_required(login_url='/authentication/login')
 def index(request):
     expenses = Expense.objects.filter(owner = request.user)
-    paginator = Paginator(expenses, 2)
+    paginator = Paginator(expenses, 5)
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
     context = {
@@ -57,7 +78,7 @@ def add_expense(request):
 
 
 def expense_edit(request, id):
-    expense= Expense.objects.get( pk = id )
+    expense = Expense.objects.get( pk = id )
     categories = Category.objects.all()
     context = {
         'expense': expense,
@@ -100,8 +121,18 @@ def expense_edit(request, id):
         messages.success(request, 'Expense Updated successfully!')
         return redirect('expenses')
 
+# def delete_expense(request, id):
+#     expense= Expense.objects.get(pk = id)
+#     expense.delete()
+#     messages.success(request, 'Expense removed!')
+#     return redirect('expenses')
+
 def delete_expense(request, id):
-    expense= Expense.objects.get(pk = id)
-    expense.delete()
-    messages.success(request, 'Expense removed!')
-    return redirect('expenses')
+    if request.method == 'POST':
+        expense = Expense.objects.get(pk=id)
+        expense.delete()
+        messages.success(request, 'Expense removed!')
+        return redirect('expenses')
+    else:
+        expense = Expense.objects.get(pk=id)
+        return render(request, 'expenses/delete_expense.html', {'expense': expense})
